use separate namespace for xxh functions (#148)
