+++ scripts/collect-coverage.sh
@@ -43,7 +43,6 @@ LCOV=(
              "$project_dir"'/examples/*' \
              "$project_dir"'/tools/*' \
              "$project_dir"'/tests/*' \
-             "$project_dir"'/include/velocypack/velocypack-xxhash*' \
              "$project_dir"'/include/velocypack/velocypack-wyhash*' \
              "$project_dir"'/src/xxhash*' \
              "$project_dir"'/src/fasthash*' \
