+++ src/xxhash.c
@@ -39,4 +39,4 @@
 #define XXH_STATIC_LINKING_ONLY /* access advanced declarations */
 #define XXH_IMPLEMENTATION      /* access definitions */
 
-#include "velocypack/velocypack-xxhash.h"
+#include "xxhash.h"
