+++ src/xxhash.h
